ItemDto追加

+++ backend/src/items/items.service.ts
@@ -1,9 +1,12 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { toDto, toDtoArray } from 'src/common/common';
 import { Repository } from 'typeorm';
 
+import { CreateItemDto } from './dto/create-item.dto';
+import { ItemDto } from './dto/item.dto';
+import { UpdateItemDto } from './dto/update-item.dto';
 import { Item } from './item.entity';
-import { Status } from './status.enum';
 
 @Injectable()
 export class ItemsService {
@@ -12,28 +15,52 @@ export class ItemsService {
     private readonly itemRepository: Repository<Item>,
   ) {}
 
-  async findAll(): Promise<Item[]> {
-    return await this.itemRepository.find();
+  async findAll(userId: number): Promise<ItemDto[]> {
+    return toDtoArray(
+      ItemDto,
+      await this.itemRepository.findBy({ userId: userId }),
+    );
   }
 
-  async findById(id: number): Promise<Item | null> {
-    return await this.itemRepository.findOneBy({ id: id });
+  async findById(id: number, userId: number): Promise<ItemDto | null> {
+    return toDto(
+      ItemDto,
+      await this.itemRepository.findOneBy({ id: id, userId: userId }),
+    );
   }
 
-  async create(item: Item): Promise<Item> {
-    return await this.itemRepository.save(item);
+  async create(userId: number, item: CreateItemDto): Promise<ItemDto> {
+    const createItem = {
+      body: item.body,
+      status: item.status,
+      userId: userId,
+    } as Item;
+    return toDto(ItemDto, await this.itemRepository.save(createItem));
   }
 
-  async updateStatus(id: number, status: Status): Promise<Item> {
-    const targetItem = await this.findById(id);
+  async update(
+    id: number,
+    userId: number,
+    item: UpdateItemDto,
+  ): Promise<ItemDto> {
+    const targetItem = await this.findById(id, userId);
     if (!targetItem) {
-      throw new NotFoundException(`Item with ID ${id} not found`);
+      throw new NotFoundException();
     }
-    targetItem.status = status;
-    return await this.itemRepository.save(targetItem);
+    // bodyとstatusのみ更新
+    const updateItem = {
+      ...targetItem,
+      body: item.body,
+      status: item.status,
+    } as Item;
+    return toDto(ItemDto, await this.itemRepository.save(updateItem));
   }
 
-  async delete(id: number): Promise<void> {
+  async delete(id: number, userId: number): Promise<void> {
+    const targetItem = await this.findById(id, userId);
+    if (!targetItem) {
+      throw new NotFoundException();
+    }
     await this.itemRepository.delete(id);
   }
 }
